+++ scripts/toggleUsePreviewBounds.js
@@ -1,9 +1,8 @@
 ﻿/* ===============================================================================================================================================
-   toggleShowHandles
+   toggleUsePreviewBounds
 
    Description
-   This script toggles the Preferences > Selection & Anchor Display > Anchor Points, Handle, and Bounding Box Display >
-   Show handles when multiple anchors are selected.
+   This script toggles the Align panel menu > Use Preview Bounds or the Preferences > General > Use Preview Bounds.
 
    Usage
    Just run this script from File > Scripts > Other Script...
@@ -16,7 +15,7 @@
    Illustrator CC or higher
 
    Version
-   1.0.1
+   1.0.0
 
    Homepage
    github.com/sky-chaser-high/adobe-illustrator-scripts
@@ -32,13 +31,11 @@
 
 
 function main() {
-    var key = 'showDirectionHandles';
+    var key = 'includeStrokeInBounds';
     var pref = app.preferences;
-    var handle = pref.getBooleanPreference(key);
-    pref.setBooleanPreference(key, !handle);
-
-    var items = (app.documents.length) ? app.activeDocument.selection : [];
-    if (!items.length) showDialog(!handle);
+    var bounds = pref.getBooleanPreference(key);
+    pref.setBooleanPreference(key, !bounds);
+    showDialog(!bounds);
 }
 
 
@@ -50,13 +47,13 @@ function isValidVersion() {
 }
 
 
-function showDialog(handle) {
+function showDialog(bounds) {
     $.localize = true;
     var ui = localizeUI();
 
     var dialog = new Window('dialog');
     dialog.text = ui.title;
-    dialog.preferredSize.width = 350;
+    dialog.preferredSize.width = 260;
     dialog.orientation = 'column';
     dialog.alignChildren = ['fill', 'top'];
     dialog.spacing = 10;
@@ -69,7 +66,7 @@ function showDialog(handle) {
     group1.margins = 20;
 
     var statictext1 = group1.add('statictext', undefined, undefined, { name: 'statictext1' });
-    statictext1.text = handle ? ui.on : ui.off;
+    statictext1.text = bounds ? ui.on : ui.off;
 
     var group2 = dialog.add('group', undefined, { name: 'group2' });
     group2.orientation = 'row';
@@ -88,8 +85,8 @@ function showDialog(handle) {
 function localizeUI() {
     return {
         title: {
-            en: 'Show handles when multiple anchors are selected',
-            ja: '複数アンカーを選択時にハンドルを表示'
+            en: 'Use Preview Bounds',
+            ja: 'プレビュー境界を使用'
         },
         on: {
             en: 'ON',